+++ client/modules/components/src/index.ts
@@ -1,4 +1,4 @@
-export { default as Sheaf } from "./Sheaf.svelte"
-export { default as SheafPanel } from "./SheafPanel.svelte"
+export { default as Sheaf } from "./sheaf/Sheaf.svelte"
+export { default as SheafPanel } from "./sheaf/SheafPanel.svelte"
 export { default as Tabview } from "./Tabview.svelte"
 export { default as IntersectionPoint } from "./IntersectionPoint.svelte"
